@@ -5,9 +5,12 @@ import time
 import espnow
 import machine
 import network
-from config import CONFIG
-from setup import setup_sensors
-from value_cache import process_sensor_data, store_sensor_data
+from config import CONFIG  # ty: ignore[unresolved-import]
+from setup import setup_sensors  # ty: ignore[unresolved-import]
+from value_cache import (  # ty: ignore[unresolved-import]
+    process_sensor_data,
+    store_sensor_data,
+)
 
 
 def run():
