@@ -1,9 +1,12 @@
 import sys
 import time
 
-from config import CONFIG
-from setup import setup_sensors
-from value_cache import process_sensor_data, store_sensor_data
+from config import CONFIG  # ty: ignore[unresolved-import]
+from setup import setup_sensors  # ty: ignore[unresolved-import]
+from value_cache import (  # ty: ignore[unresolved-import]
+    process_sensor_data,
+    store_sensor_data,
+)
 
 
 def run():
@@ -23,6 +26,6 @@ while True:
         run()
     except Exception as exc:
         print("error running sensor:")
-        sys.print_exception(exc)  # type: ignore[attr-defined]
+        sys.print_exception(exc)  # ty: ignore[unresolved-attribute]
         # Wait for a 10th of an interval.
         time.sleep(CONFIG["interval"] / 10)
